2024 12 23
1. .gitignore 수정
2. lambda, eventbridge 구성 및 배포
3. lambda에서 실행할 수 있도록 수정
