+++ .gitignore
@@ -146,3 +146,8 @@ cython_debug/
 
 # Log file
 discord.log
+
+# AWS Lambda
+deployment.zip
+lambda_scraper/*.zip
+lambda_scraper/distribution/
