+++ README.md
@@ -1,136 +1,49 @@
-# Python Discord Bot Template
+# SAS (SSU Algorithm Solver) Bot
 
-<p align="center">
-  <a href="https://discord.gg/mTBrXyWxAF"><img src="https://img.shields.io/discord/739934735387721768?logo=discord"></a>
-  <a href="https://github.com/kkrypt0nn/Python-Discord-Bot-Template/releases"><img src="https://img.shields.io/github/v/release/kkrypt0nn/Python-Discord-Bot-Template"></a>
-  <a href="https://github.com/kkrypt0nn/Python-Discord-Bot-Template/commits/main"><img src="https://img.shields.io/github/last-commit/kkrypt0nn/Python-Discord-Bot-Template"></a>
-  <a href="https://github.com/kkrypt0nn/Python-Discord-Bot-Template/blob/main/LICENSE.md"><img src="https://img.shields.io/github/license/kkrypt0nn/Python-Discord-Bot-Template"></a>
-  <a href="https://github.com/kkrypt0nn/Python-Discord-Bot-Template"><img src="https://img.shields.io/github/languages/code-size/kkrypt0nn/Python-Discord-Bot-Template"></a>
-  <a href="https://conventionalcommits.org/en/v1.0.0/"><img src="https://img.shields.io/badge/Conventional%20Commits-1.0.0-%23FE5196?logo=conventionalcommits&logoColor=white"></a>
-  <a href="https://github.com/psf/black"><img src="https://img.shields.io/badge/code%20style-black-000000.svg"></a>
-</p>
+이 프로젝트는 [Python Discord Bot Template](https://github.com/kkrypt0nn/Python-Discord-Bot-Template)을 기반으로 개발되었습니다.
 
-This repository is a template that everyone can use for the start of their Discord bot.
+## 프로젝트 소개
+BOJ(Baekjoon Online Judge) 알고리즘 문제 풀이 현황을 디스코드를 통해 확인할 수 있는 봇입니다.
 
-When I first started creating my Discord bot it took me a while to get everything setup and working with cogs and more.
-I would've been happy if there were any template existing. However, there wasn't any existing template. That's why I
-decided to create my own template to let **you** guys create your Discord bot easily.
+## 주요 기능
+- 매시간 BOJ 제출 현황 자동 수집
+- 디스코드를 통한 문제 풀이 현황 확인
+- solved.ac 정보 확인
 
-Please note that this template is not supposed to be the best template, but a good template to start learning how
-discord.py works and to make your own bot easily.
-
-If you plan to use this template to make your own template or bot, you **have to**:
-
-- Keep the credits, and a link to this repository in all the files that contains my code
-- Keep the same license for unchanged code
-
-See [the license file](https://github.com/kkrypt0nn/Python-Discord-Bot-Template/blob/master/LICENSE.md) for more
-information, I reserve the right to take down any repository that does not meet these requirements.
-
-## Support
-
-Before requesting support, you should know that this template requires you to have at least a **basic knowledge** of
-Python and the library is made for **advanced users**. Do not use this template if you don't know the
-basics or some advanced topics such as OOP or async. [Here's](https://pythondiscord.com/pages/resources) a link for resources to learn python.
-
-If you need some help for something, do not hesitate to join my Discord server [here](https://discord.gg/mTBrXyWxAF).
-
-All the updates of the template are available [here](UPDATES.md).
-
-## Disclaimer
-
-Slash commands can take some time to get registered globally, so if you want to test a command you should use
-the `@app_commands.guilds()` decorator so that it gets registered instantly. Example:
-
-```py
-@commands.hybrid_command(
-  name="command",
-  description="Command description",
-)
-@app_commands.guilds(discord.Object(id=GUILD_ID)) # Place your guild ID here
-```
-
-When using the template you confirm that you have read the [license](LICENSE.md) and comprehend that I can take down
-your repository if you do not meet these requirements.
-
-## How to download it
-
-This repository is now a template, on the top left you can simply click on "**Use this template**" to create a GitHub
-repository based on this template.
-
-Alternatively you can do the following:
-
-- Clone/Download the repository
-  - To clone it and get the updates you can definitely use the command
-    `git clone`
-- Create a Discord bot [here](https://discord.com/developers/applications)
-- Get your bot token
-- Invite your bot on servers using the following invite:
-  https://discord.com/oauth2/authorize?&client_id=YOUR_APPLICATION_ID_HERE&scope=bot+applications.commands&permissions=PERMISSIONS (
-  Replace `YOUR_APPLICATION_ID_HERE` with the application ID and replace `PERMISSIONS` with the required permissions
-  your bot needs that it can be get at the bottom of a this
-  page https://discord.com/developers/applications/YOUR_APPLICATION_ID_HERE/bot)
-
-## How to set up
-
-To set up the bot it was made as simple as possible.
-
-### `config.json` file
-
-There is [`config.json`](config.json) file where you can put the
-needed things to edit.
-
-Here is an explanation of what everything is:
-
-| Variable                  | What it is                                     |
-| ------------------------- | ---------------------------------------------- |
-| YOUR_BOT_PREFIX_HERE      | The prefix you want to use for normal commands |
-| YOUR_BOT_INVITE_LINK_HERE | The link to invite the bot                     |
-
-### `.env` file
-
-To set up the token you will have to make use of the [`.env.example`](.env.example) file, you should rename it to `.env` and replace `YOUR_BOT_TOKEN_HERE` with your actual bot's token.
-
-Alternatively you can simply create an environment variable named `TOKEN`.
-
-## How to start
-
-To start the bot you simply need to launch, either your terminal (Linux, Mac & Windows), or your Command Prompt (
-Windows)
+## 프로젝트 구조
+```tree
 .
-
-Before running the bot you will need to install all the requirements with this command:
-
+├── LICENSE.md
+├── README.md
+├── bot.py
+├── cogs
+│   ├── fun.py
+│   ├── general.py
+│   ├── information.py
+│   ├── moderation.py
+│   ├── owner.py
+│   ├── register.py
+│   ├── template.py
+│   └── today.py
+├── config.json # 직접 추가해야함
+├── database
+│   ├── __init__.py
+│   ├── __pycache__
+│   └── dynamodb.py
+├── lambda_scraper
+│   ├── __init__.py
+│   ├── lambda_database.py
+│   ├── lambda_function.py
+│   ├── readme.md
+│   ├── requirements.txt
+│   └── scraper.py
+├── poetry.lock
+├── pyproject.toml
+└── response
+    ├── __init__.py
+    └── bot_response_object.py
 ```
-python -m pip install -r requirements.txt
-```
 
-After that you can start it with
-
-```
-python bot.py
-```
-
-> **Note** You may need to replace `python` with `py`, `python3`, `python3.11`, etc. depending on what Python versions you have installed on the machine.
-
-## Issues or Questions
-
-If you have any issues or questions of how to code a specific command, you can:
-
-- Join my Discord server [here](https://discord.gg/mTBrXyWxAF)
-- Post them [here](https://github.com/kkrypt0nn/Python-Discord-Bot-Template/issues)
-
-Me or other people will take their time to answer and help you.
-
-## Versioning
-
-We use [SemVer](http://semver.org) for versioning. For the versions available, see
-the [tags on this repository](https://github.com/kkrypt0nn/Python-Discord-Bot-Template/tags).
-
-## Built With
-
-- [Python 3.11.5](https://www.python.org/)
-
-## License
-
-This project is licensed under the Apache License 2.0 - see the [LICENSE.md](LICENSE.md) file for details
+## 라이선스
+이 프로젝트는 Apache License 2.0 라이선스를 따르고 있습니다.
+자세한 내용은 [LICENSE.md](LICENSE.md) 파일을 참조해주세요.
