+++ lambda_scraper/requirements.txt
@@ -1,0 +1,4 @@
+aiohttp
+boto3
+beautifulsoup4
+python-dotenv
